mycart/requirement.txt
asgiref==3.5.2
Django==4.1.1
django-cors-headers==3.13.0
django-dotenv==1.4.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.1
Pillow==9.2.0
PyJWT==2.5.0
pytz==2022.4
sqlparse==0.4.3
tzdata==2022.2
